backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
psycopg2-binary
httpx
beautifulsoup4